requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
google-genai>=0.3.0
python-dotenv>=1.0.1
pydantic>=2.6.0
pydantic-settings>=2.2.0
httpx>=0.27.0
Pillow>=10.3.0
aiofiles>=23.2.1
slowapi>=0.1.9
streamlit>=1.40.0
